trading-bots-and-scripts/cointegration/requirements.txt
python-dotenv==0.20.0
pandas==1.5.1
statsmodels==0.13.5
matplotlib==3.6.2
pybit==2.4.1
